+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [#38](https://github.com/tweag/genealogos/pull/38) display nixtract's status information when running
 - [#44](https://github.com/tweag/genealogos/pull/44) adds two functions to the `Backend` trait to set options
 - [#48](https://github.com/tweag/genealogos/pull/48) [#49](https://github.com/tweag/genealogos/pull/49) added a NixOS module to our flake
+- [#50](https://github.com/tweag/genealogos/pull/50) included the frontend in rocket behind the `frontend` feature flag
 
 ### Changed
 - [#41](https://github.com/tweag/genealogos/pull/41) reworked the Genealogos fronend, paving the way for supporting other bom formats